Moved caching of groups
